[#72] Move/adapt tests
Tests related to MIME type validation are moved to
the file component validation tests, and simplified a bit.
