+++ src/openforms/formio/tests/files/image-256x256.pdf
