+++ src/openforms/submissions/tests/test_submission_attachment.py
@@ -8,11 +8,9 @@ from django.urls import reverse
 from maykin_2fa.test import disable_admin_mfa
 from PIL import Image, UnidentifiedImageError
 from privates.test import temp_private_root
-from rest_framework.exceptions import ValidationError
 
 from openforms.accounts.tests.factories import SuperUserFactory
 from openforms.api.exceptions import RequestEntityTooLarge
-from openforms.config.models import GlobalConfiguration
 from openforms.forms.tests.factories import FormStepFactory
 
 from ..attachments import (
@@ -22,7 +20,6 @@ from ..attachments import (
     cleanup_submission_temporary_uploaded_files,
     resize_attachment,
     resolve_uploads_from_data,
-    validate_uploads,
 )
 from ..models import SubmissionFileAttachment
 from .factories import (
@@ -1096,335 +1093,6 @@ class SubmissionAttachmentTest(TestCase):
 
         with self.assertRaises(RequestEntityTooLarge):
             attach_uploads_to_submission_step(submission_step)
-
-    @tag("GHSA-h85r-xv4w-cg8g")
-    def test_attach_upload_validates_file_content_types_malicious_content(self):
-        """
-        Regression test for CVE-2022-31041 to ensure the file content is validated
-        against the formio configuration.
-
-        We cannot rely on file extension or browser mime-type. Therefore, we have a test
-        file that claims to be a PDF but is actually an image that we put in the upload
-        data. The step attaching the uploads to the form data must validate the
-        configuration.
-        """
-        with open(TEST_FILES_DIR / "image-256x256.pdf", "rb") as infile:
-            upload1 = TemporaryFileUploadFactory.create(
-                file_name="my-pdf.pdf",
-                content=File(infile),
-                content_type="application/pdf",
-            )
-            upload2 = TemporaryFileUploadFactory.create(
-                file_name="my-pdf2.pdf", content=File(infile), content_type="image/png"
-            )
-
-        data = {
-            "my_file": [
-                {
-                    "url": f"http://server/api/v2/submissions/files/{upload1.uuid}",
-                    "data": {
-                        "url": f"http://server/api/v2/submissions/files/{upload1.uuid}",
-                        "form": "",
-                        "name": "my-pdf.pdf",
-                        "size": 585,
-                        "baseUrl": "http://server",
-                        "project": "",
-                    },
-                    "name": "my-pdf-12305610-2da4-4694-a341-ccb919c3d543.png",
-                    "size": 585,
-                    "type": "application/pdf",  # we are lying!
-                    "storage": "url",
-                    "originalName": "my-pdf.pdf",
-                },
-                {
-                    "url": f"http://server/api/v2/submissions/files/{upload2.uuid}",
-                    "data": {
-                        "url": f"http://server/api/v2/submissions/files/{upload2.uuid}",
-                        "form": "",
-                        "name": "my-pdf2.pdf",
-                        "size": 585,
-                        "baseUrl": "http://server",
-                        "project": "",
-                    },
-                    "name": "my-pdf2-12305610-2da4-4694-a341-ccb919c3d543.png",
-                    "size": 585,
-                    "type": "image/png",  # we are lying!
-                    "storage": "url",
-                    "originalName": "my-pdf2.pdf",
-                },
-            ],
-        }
-        formio_components = {
-            "key": "my_file",
-            "type": "file",
-            "multiple": True,
-            "file": {
-                "name": "",
-                "type": ["application/pdf"],
-            },
-            "filePattern": "application/pdf",
-        }
-
-        submission = SubmissionFactory.from_components(
-            [formio_components],
-            submitted_data=data,
-        )
-        submission_step = submission.submissionstep_set.get()
-
-        with self.assertRaises(ValidationError) as err_context:
-            validate_uploads(submission_step, data=data)
-
-        validation_error = err_context.exception.get_full_details()
-        self.assertEqual(len(validation_error["my_file"]), 2)
-
-    @tag("GHSA-h85r-xv4w-cg8g")
-    def test_attach_upload_validates_file_content_types_ok(self):
-        """
-        Regression test for CVE-2022-31041 to ensure the file content is validated
-        against the formio configuration.
-
-        We cannot rely on file extension or browser mime-type. Therefore, we have a test
-        file that claims to be a PDF but is actually an image that we put in the upload
-        data. The step attaching the uploads to the form data must validate the
-        configuration.
-        """
-        with open(TEST_FILES_DIR / "image-256x256.png", "rb") as infile:
-            upload1 = TemporaryFileUploadFactory.create(
-                file_name="my-img.png",
-                content=File(infile),
-                content_type="image/png",
-            )
-            upload2 = TemporaryFileUploadFactory.create(
-                file_name="my-img2.png", content=File(infile), content_type="image/png"
-            )
-
-        data = {
-            "my_file": [
-                {
-                    "url": f"http://server/api/v2/submissions/files/{upload1.uuid}",
-                    "data": {
-                        "url": f"http://server/api/v2/submissions/files/{upload1.uuid}",
-                        "form": "",
-                        "name": "my-img.png",
-                        "size": 585,
-                        "baseUrl": "http://server",
-                        "project": "",
-                    },
-                    "name": "my-img-12305610-2da4-4694-a341-ccb919c3d543.png",
-                    "size": 585,
-                    "type": "image/png",  # we are lying!
-                    "storage": "url",
-                    "originalName": "my-img.png",
-                },
-                {
-                    "url": f"http://server/api/v2/submissions/files/{upload2.uuid}",
-                    "data": {
-                        "url": f"http://server/api/v2/submissions/files/{upload2.uuid}",
-                        "form": "",
-                        "name": "my-img2.png",
-                        "size": 585,
-                        "baseUrl": "http://server",
-                        "project": "",
-                    },
-                    "name": "my-img2-12305610-2da4-4694-a341-ccb919c3d543.png",
-                    "size": 585,
-                    "type": "image/png",  # we are lying!
-                    "storage": "url",
-                    "originalName": "my-img2.png",
-                },
-            ],
-        }
-        formio_components = {
-            "key": "my_file",
-            "type": "file",
-            "multiple": True,
-            "file": {
-                "name": "",
-                "type": ["image/png", "image/jpeg"],
-            },
-            "filePattern": "image/png,image/jpeg",
-        }
-
-        submission = SubmissionFactory.from_components(
-            [formio_components],
-            submitted_data=data,
-        )
-        submission_step = submission.submissionstep_set.get()
-
-        try:
-            validate_uploads(submission_step, data=data)
-        except ValidationError:
-            self.fail("Uploads should be accepted since the content types are valid")
-
-    @tag("GHSA-h85r-xv4w-cg8g")
-    def test_attach_upload_validates_file_content_types_implicit_wildcard(self):
-        with open(TEST_FILES_DIR / "image-256x256.png", "rb") as infile:
-            upload = TemporaryFileUploadFactory.create(
-                file_name="my-img.png",
-                content=File(infile),
-                content_type="image/png",
-            )
-
-        data = {
-            "my_file": [
-                {
-                    "url": f"http://server/api/v2/submissions/files/{upload.uuid}",
-                    "data": {
-                        "url": f"http://server/api/v2/submissions/files/{upload.uuid}",
-                        "form": "",
-                        "name": "my-img.png",
-                        "size": 585,
-                        "baseUrl": "http://server",
-                        "project": "",
-                    },
-                    "name": "my-img-12305610-2da4-4694-a341-ccb919c3d543.png",
-                    "size": 585,
-                    "type": "image/png",  # we are lying!
-                    "storage": "url",
-                    "originalName": "my-img.png",
-                },
-            ],
-        }
-        formio_components = {
-            "key": "my_file",
-            "type": "file",
-            "multiple": True,
-            "file": {
-                "name": "",
-            },
-            "filePattern": "",
-        }
-
-        submission = SubmissionFactory.from_components(
-            [formio_components],
-            submitted_data=data,
-        )
-        submission_step = submission.submissionstep_set.get()
-
-        try:
-            validate_uploads(submission_step, data=data)
-        except ValidationError:
-            self.fail(
-                "Uploads should be accepted since the content types match the wildcard"
-            )
-
-    @tag("GHSA-h85r-xv4w-cg8g")
-    def test_attach_upload_validates_file_content_types_wildcard(self):
-        """
-        Regression test for the initial CVE-2022-31041 patch.
-
-        Assert that file uploads are allowed if the "All" file types in the file
-        configuration tab is used, which presents as a '*' entry.
-        """
-        with open(TEST_FILES_DIR / "image-256x256.png", "rb") as infile:
-            upload = TemporaryFileUploadFactory.create(
-                file_name="my-img.png",
-                content=File(infile),
-                content_type="image/png",
-            )
-
-        data = {
-            "my_file": [
-                {
-                    "url": f"http://server/api/v2/submissions/files/{upload.uuid}",
-                    "data": {
-                        "url": f"http://server/api/v2/submissions/files/{upload.uuid}",
-                        "form": "",
-                        "name": "my-img.png",
-                        "size": 585,
-                        "baseUrl": "http://server",
-                        "project": "",
-                    },
-                    "name": "my-img-12305610-2da4-4694-a341-ccb919c3d543.png",
-                    "size": 585,
-                    "type": "image/png",  # we are lying!
-                    "storage": "url",
-                    "originalName": "my-img.png",
-                },
-            ],
-        }
-        formio_components = {
-            "key": "my_file",
-            "type": "file",
-            "multiple": True,
-            "file": {
-                "name": "",
-                "type": ["*"],
-            },
-            "filePattern": "*",
-        }
-
-        submission = SubmissionFactory.from_components(
-            [formio_components],
-            submitted_data=data,
-        )
-        submission_step = submission.submissionstep_set.get()
-
-        try:
-            validate_uploads(submission_step, data=data)
-        except ValidationError:
-            self.fail(
-                "Uploads should be accepted since the content types match the wildcard"
-            )
-
-    @patch("openforms.submissions.attachments.GlobalConfiguration.get_solo")
-    def test_attach_upload_validates_file_content_types_default_configuration(
-        self, m_solo
-    ):
-        m_solo.return_value = GlobalConfiguration(
-            form_upload_default_file_types=["application/pdf", "image/jpeg"],
-        )
-
-        with open(TEST_FILES_DIR / "image-256x256.png", "rb") as infile:
-            upload = TemporaryFileUploadFactory.create(
-                file_name="my-img.png",
-                content=File(infile),
-                content_type="image/png",
-            )
-
-        data = {
-            "my_file": [
-                {
-                    "url": f"http://server/api/v2/submissions/files/{upload.uuid}",
-                    "data": {
-                        "url": f"http://server/api/v2/submissions/files/{upload.uuid}",
-                        "form": "",
-                        "name": "my-img.png",
-                        "size": 585,
-                        "baseUrl": "http://server",
-                        "project": "",
-                    },
-                    "name": "my-img-12305610-2da4-4694-a341-ccb919c3d543.png",
-                    "size": 585,
-                    "type": "image/png",  # we are lying!
-                    "storage": "url",
-                    "originalName": "my-img.png",
-                },
-            ],
-        }
-        formio_components = {
-            "key": "my_file",
-            "type": "file",
-            "multiple": True,
-            "file": {
-                "name": "",
-                "type": ["*"],
-            },
-            "filePattern": "*",
-            "useConfigFiletypes": True,
-        }
-
-        submission = SubmissionFactory.from_components(
-            [formio_components],
-            submitted_data=data,
-        )
-        submission_step = submission.submissionstep_set.get()
-
-        with self.assertRaises(ValidationError) as err_context:
-            validate_uploads(submission_step, data=data)
-
-        validation_error = err_context.exception.get_full_details()
-        self.assertEqual(len(validation_error["my_file"]), 1)
 
     @disable_admin_mfa()
     def test_attachment_retrieve_view_requires_permission(self):
